Move engineering functions to sub package
